Fix issues with lowercased model names
