+++ Package.resolved
@@ -6,8 +6,8 @@
         "repositoryURL": "https://github.com/lunarway/SwaggerSwiftML",
         "state": {
           "branch": null,
-          "revision": "a4cb6443f122a67181d9200eabd663bf49d11c78",
-          "version": "1.0.10"
+          "revision": "28607323c29b474e879075ef43f54892327151cc",
+          "version": "1.0.11"
         }
       },
       {
