+++ Sources/SwaggerSwift/Extensions/Schema.swift
@@ -1,0 +1,19 @@
+import SwaggerSwiftML
+
+extension Schema {
+    var isInternalOnly: Bool {
+        if let value = self.customFields["x-internal"] {
+            return value == "true"
+        } else {
+            return false
+        }
+    }
+
+    var overridesName: String? {
+        if let value = self.customFields["x-override-name"] {
+            return value
+        } else {
+            return nil
+        }
+    }
+}
