+++ Sources/SwaggerSwift/Models/TypeType.swift
@@ -41,293 +41,3 @@ indirect enum TypeType {
         }(self) + (required ? "" : "?")
     }
 }
-
-func getType(forSchema schema: SwaggerSwiftML.Schema, typeNamePrefix: String, swagger: Swagger) -> (TypeType, [ModelDefinition]) {
-    switch schema.type {
-    case .string(format: let format, let enumValues, _, _, _):
-        if let enumValues = enumValues {
-            let enumTypename = typeNamePrefix
-            let def = ModelDefinition.enumeration(Enumeration(serviceName: swagger.serviceName,
-                                                              description: schema.description,
-                                                              typeName: enumTypename,
-                                                              values: enumValues,
-                                                              isCodable: true))
-            return (.object(typeName: enumTypename), [def])
-        }
-
-        if let format = format {
-            switch format {
-            case .string:
-                return (.string, [])
-            case .date:
-                return (.object(typeName: "Date"), [])
-            case .dateTime:
-                return (.object(typeName: "Date"), [])
-            case .password:
-                return (.string, [])
-            case .email:
-                return (.string, [])
-            case .binary:
-                return (.object(typeName: "Data"), [])
-            case .long: fallthrough
-            case .float: fallthrough
-            case .double: fallthrough
-            case .byte: fallthrough
-            case .boolean: fallthrough
-            case .int32:
-                fatalError("Found unsupported field")
-            case .unsupported(let unsupported):
-                switch unsupported {
-                case "ISO8601":
-                return (.date, [])
-                case "uuid":
-                    return (.object(typeName: "String"), [])
-                default:
-                    fatalError("Found unsupported field: \(unsupported)")
-                }
-            }
-        } else {
-            return (.string, [])
-        }
-    case .integer(let format,_,_,_,_,_):
-        if let format = format {
-            switch format {
-            case .long:
-                return (.double, [])
-            case .float: return (.float, [])
-            case .int32: return (.int, [])
-            case .double: return (.double, [])
-            case .date: fallthrough
-            case .dateTime: fallthrough
-            case .password: fallthrough
-            case .email: fallthrough
-            case .string: fallthrough
-            case .byte: fallthrough
-            case .binary: fallthrough
-            case .boolean:
-                fatalError("This cannot happen for this case")
-            case .unsupported(let unsupported):
-                switch unsupported {
-                case "int":
-                    return (.int, [])
-                case "int64":
-                    return (.int64, [])
-                case "decimal":
-                    return (.double, [])
-                default:
-                    fatalError("Unsupported field type received: \(unsupported)")
-                }
-            }
-        }
-
-        return (.int, [])
-    case .number(let format, _, _, _, _, _):
-        if let format = format {
-            switch format {
-            case .long:
-                return (.double, [])
-            case .float: return (.float, [])
-            case .int32: return (.int, [])
-            case .double: return (.double, [])
-            case .date: fallthrough
-            case .dateTime: fallthrough
-            case .password: fallthrough
-            case .email: fallthrough
-            case .string: fallthrough
-            case .byte: fallthrough
-            case .binary: fallthrough
-            case .boolean:
-                fatalError("This cannot happen for this case")
-            case .unsupported(let unsupported):
-                switch unsupported {
-                case "int", "integer":
-                    return (.int, [])
-                case "int64":
-                    return (.int64, [])
-                case "decimal":
-                    return (.double, [])
-                default:
-                    fatalError("Unsupported field type received: \(unsupported)")
-                }
-            }
-        }
-
-        return (.double, [])
-    case .boolean:
-        return (.boolean, [])
-    case .array(let items, _, _, _, _):
-        let type = typeOfItems(schema: schema,
-                               items: items,
-                               typeNamePrefix: "\(typeNamePrefix)Item",
-                               swagger: swagger)
-        return (.array(typeName: type.0), type.1)
-    case .object(let properties, allOf: let allOf):
-        return parseObject(required: [],
-                           properties: properties, allOf: allOf, swagger: swagger, typeNamePrefix: typeNamePrefix, schema: schema, customFields: [:])
-    case .dictionary(valueType: let valueType, keys: _):
-        switch valueType {
-        case .any:
-            return (.object(typeName: "AdditionalProperty"), [])
-        case .reference(_):
-            fatalError("not supported")
-        case .schema(_):
-            fatalError("not supported")
-        }
-    case .file:
-        return (.void, [])
-    }
-}
-
-private func typeOfItems(schema: Schema, items: Node<Items>, typeNamePrefix: String, swagger: Swagger) -> (TypeType, [ModelDefinition]) {
-    switch items {
-    case .reference(let ref):
-        let schema = swagger.findSchema(node: .reference(ref))
-        if case SchemaType.object = schema.type {
-            return (.object(typeName: ref.components(separatedBy: "/").last!), [])
-        } else {
-            fatalError()
-        }
-    case .node(let node):
-        switch node.type {
-        case .string(format: _, enumValues: _, maxLength: _, minLength: _, pattern: _):
-            return (.string, [])
-        case .number(format: _, maximum: _, exclusiveMaximum: _, minimum: _, exclusiveMinimum: _, multipleOf: _):
-            return (.int, [])
-        case .integer(format: _, maximum: _, exclusiveMaximum: _, minimum: _, exclusiveMinimum: _, multipleOf: _):
-            return (.int, [])
-        case .boolean:
-            return (.boolean, [])
-        case .array(let items, collectionFormat: _, maxItems: _, minItems: _, uniqueItems: _):
-            return typeOfItems(schema: schema, items: Node.node(items), typeNamePrefix: typeNamePrefix, swagger: swagger)
-        case .object(required: let required, properties: let properties, allOf: let allOf):
-            return parseObject(required: required,
-                               properties: properties,
-                               allOf: allOf,
-                               swagger: swagger,
-                               typeNamePrefix: typeNamePrefix,
-                               schema: schema,
-                               customFields: node.customFields)
-        }
-    }
-}
-
-extension Schema {
-    var isInternalOnly: Bool {
-        if let value = self.customFields["x-internal"] {
-            return value == "true"
-        } else {
-            return false
-        }
-    }
-
-    var overridesName: String? {
-        if let value = self.customFields["x-override-name"] {
-            return value
-        } else {
-            return nil
-        }
-    }
-}
-
-extension SwaggerSwiftML.Operation {
-    var isInternalOnly: Bool {
-        if let value = self.customFields["x-internal"] {
-            return value == "true"
-        } else {
-            return false
-        }
-    }
-}
-
-func parseObject(required: [String], properties: [String: Node<Schema>], allOf: [Node<Schema>]?, swagger: Swagger, typeNamePrefix: String, schema: Schema, customFields: [String: String]) -> (TypeType, [ModelDefinition]) {
-    let typeName = (customFields["x-override-name"] ?? schema.overridesName ?? typeNamePrefix).uppercasingFirst
-
-    if let allOf = allOf {
-        let result: [(String?, [ModelField], [ModelDefinition])] = allOf.map {
-            switch $0 {
-            case .reference(let reference):
-                let node = swagger.findSchema(node: .reference(reference))
-                if case SchemaType.object = node.type {
-                    let typeName = reference.components(separatedBy: "/").last ?? ""
-                    return (typeName, [], [])
-                } else {
-                    let result = getType(forSchema: node, typeNamePrefix: typeName, swagger: swagger)
-                    return (nil, [], result.1)
-                }
-            case .node(let innerSchema):
-                if case let SchemaType.object(properties, allOf) = innerSchema.type {
-                    assert(allOf == nil, "Not implemented")
-                    let result: [(ModelField, [ModelDefinition])] = properties.map {
-                        switch $0.value {
-                        case .reference(let reference):
-                            let node = swagger.findSchema(node: .reference(reference))
-                            let typeName = reference.components(separatedBy: "/").last ?? ""
-                            if case SchemaType.object = node.type {
-                                return (ModelField(description: node.description, type: .object(typeName: typeName), name: $0.key, required: innerSchema.required.contains($0.key)), [])
-                            } else {
-                                let type = getType(forSchema: node, typeNamePrefix: typeName, swagger: swagger)
-                                return (ModelField(description: node.description, type: type.0, name: $0.key, required: innerSchema.required.contains($0.key)), type.1)
-                            }
-                        case .node(let schema):
-                            let type = getType(forSchema: schema, typeNamePrefix: typeName, swagger: swagger)
-                            return (ModelField(description: schema.description, type: type.0, name: $0.key, required: innerSchema.required.contains($0.key)), type.1)
-                        }
-                    }
-
-                    return (nil, result.map { $0.0 }, result.flatMap { $0.1 })
-                } else {
-                    fatalError("Not implemented")
-                }
-            }
-        }
-
-//        let inherits = result.compactMap { $0.0 }
-
-        let models = result.flatMap { $0.2 }
-
-        let model = Model(serviceName: swagger.serviceName,
-                          description: schema.description,
-                          typeName: typeName,
-                          fields: result.flatMap { $0.1 },
-                          inheritsFrom: ["Codable"],//inherits,
-                          isInternalOnly: schema.isInternalOnly,
-                          embeddedDefinitions: models)
-
-        return (.object(typeName: typeName), [.model(model)])
-    }
-
-    let result: [(ModelField, [ModelDefinition])] = properties.map {
-        switch $0.value {
-        case .reference(let reference):
-            let node = swagger.findSchema(node: .reference(reference))
-            let typeName = reference.components(separatedBy: "/").last ?? ""
-            if case SchemaType.object = node.type {
-                return (ModelField(description: node.description, type: .object(typeName: typeName), name: $0.key, required: schema.required.contains($0.key)), [])
-            } else {
-                let type = getType(forSchema: node, typeNamePrefix: typeName, swagger: swagger)
-                return (ModelField(description: node.description, type: type.0, name: $0.key, required: schema.required.contains($0.key)), type.1)
-            }
-        case .node(let innerSchema):
-            var typeName = "\($0.key.uppercasingFirst)"
-            if typeName == "Type" {
-                typeName = "\(typeName)\($0.key.uppercasingFirst)"
-            }
-            let type = getType(forSchema: innerSchema, typeNamePrefix: typeName, swagger: swagger)
-            let modelField = ModelField(description: innerSchema.description, type: type.0, name: $0.key, required: required.contains($0.key) || schema.required.contains($0.key))
-            return (modelField, type.1)
-        }
-    }
-
-    let fields = result.map { $0.0 }
-    let inlineModels = result.flatMap { $0.1 }
-
-    let model = Model(serviceName: swagger.serviceName,
-                      description: schema.description,
-                      typeName: typeName,
-                      fields: fields.sorted(by: { $0.name < $1.name }),
-                      inheritsFrom: ["Codable"],
-                      isInternalOnly: schema.isInternalOnly,
-                      embeddedDefinitions: inlineModels)
-
-    return (.object(typeName: typeName), [.model(model)])
-}
